perfrunbook/utilities: Enable stack-reversed flamegraph to be produced
Update capture_flamegraphs.sh to take a -R or --reverse flag to produce
stack reversed flamegraphs.  Stack reversed flamegraphs aggregate by
leaf-function on the bottom taking the highest fraction of time, and
the flames are all the stack paths taken to get to those leaf functions.
This is useful when looking for optimization opportunities at the
function level.